backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
python-dotenv==1.0.1